# web/doc/swagger.yaml
openapi: 3.0.0
info:
  title: 'Accencture API'
  contact:
    email: vlad.veresklya@yandex.ru
  version: 1.0.0
paths:
...
  /api/shift:
    get:
      tags:
        - Смены
      summary: 'Найти все смены'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/shift/{id}':
    get:
      tags:
        - Смены
      summary: 'Найти смены по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID смены'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - Смены
      summary: 'Обновить смену'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                date_start:
                  description: 'Дата начала'
                  type: date
                date_end:
                  description: 'Дата конца'
                  type: date
                orders:
                  description: Заказы
                  type: json
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Смены
      summary: 'Удалить смену по id'
      operationId: f880dd98b34a74c15b96d0f358dd6ef7
      parameters:
        -
          name: id
          in: path
          description: 'ID смены'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/shift/:
    post:
      tags:
        - Смены
      summary: 'Добавить смену'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                date_start:
                  description: 'Дата начала'
                  type: date
                date_end:
                  description: 'Дата конца'
                  type: date
                orders:
                  description: Заказы
                  type: json
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
...
  /api/vagon:
    get:
      tags:
        - Вагоны
      summary: 'Найти все вагоны'
      parameters:
        -
          name: per-page
          in: query
          description: 'Кол-во записей на странице'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
  '/api/vagon/{id}':
    get:
      tags:
        - Вагоны
      summary: 'Найти вагоны по id'
      parameters:
        -
          name: id
          in: path
          description: 'ID вагона'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    put:
      tags:
        - Вагоны
      summary: 'Обновить вагон'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                auto_number:
                  description: 'Номер вагона'
                  type: string
                weight:
                  description: Вес
                  type: string
                status:
                  description: Статус
                  type: string
                fixation:
                  description: 'Фиксированные данные'
                  type: json
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
                declared_rejection_rate:
                  description: 'Заявленный процент брака (дельта)'
                  type: integer
                actual_scrap_rate:
                  description: 'Фактический процент брака ( дельта факт)'
                  type: integer
                defect:
                  description: 'Брак в тоннах '
                  type: string
                pure_material:
                  description: 'Чистое сырьё без брака'
                  type: string
                id_provider:
                  description: 'Id поставщика'
                  type: integer
                id_reception_point:
                  description: 'Id пункт приёма'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Вагоны
      summary: 'Удалить вагон по id'
      operationId: 023e589bb3fa9bf4051ddffff66b3d6f
      parameters:
        -
          name: id
          in: path
          description: 'ID вагона'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
  /api/vagon/:
    post:
      tags:
        - Вагоны
      summary: 'Добавить вагон'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                auto_number:
                  description: 'Номер вагона'
                  type: string
                weight:
                  description: Вес
                  type: string
                status:
                  description: Статус
                  type: string
                fixation:
                  description: 'Фиксированные данные'
                  type: json
                id_technologist:
                  description: 'Id технолога'
                  type: integer
                id_technician:
                  description: 'Id техника'
                  type: integer
                declared_rejection_rate:
                  description: 'Заявленный процент брака (дельта)'
                  type: integer
                actual_scrap_rate:
                  description: 'Фактический процент брака ( дельта факт)'
                  type: integer
                defect:
                  description: 'Брак в тоннах '
                  type: string
                pure_material:
                  description: 'Чистое сырьё без брака'
                  type: string
                id_provider:
                  description: 'Id поставщика'
                  type: integer
                id_reception_point:
                  description: 'Id пункт приёма'
                  type: integer
              type: object
      responses:
        '200':
          description: ''
...